requirements.txt
aiodns==3.1.1
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==4.2.0
argcomplete==3.1.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Babel==2.10.3
blinker==1.7.0
cachetools==5.5.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.7
distro==1.9.0
Flask==3.0.2
frozenlist==1.4.0
google-api-core==2.19.1
google-auth==2.34.0
google-cloud-speech==2.27.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.63.2
grpcio==1.65.5
grpcio-status==1.65.5
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.4
openai==1.12.0
packaging==24.0
pillow==10.4.0
pipx==1.4.3
platformdirs==4.2.0
proto-plus==1.24.0
protobuf==5.27.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyasyncore==1.0.2
pycares==4.4.0
pydantic==1.10.14
Pygments==2.17.2
pyinotify==0.9.6
PyJWT==2.7.0
pyOpenSSL==23.2.0
pytesseract==0.3.13
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
rich==13.7.1
rsa==4.9
setuptools==68.1.2
simplejson==3.19.2
sniffio==1.3.0
tqdm==4.66.5
twilio==8.12.0
typing_extensions==4.10.0
urllib3==2.0.7
userpath==1.9.1
Werkzeug==3.0.1
wheel==0.42.0
yarl==1.8.2
